Dedupe the e2e test rig into tests/common
`e2e_consensus.rs` carried its own copy of the ~1200-line integration rig
(`AppShardHarness`, `Tier2ArchiveRig`, `StubFrameProver`, `TestProver`,
`MaterializeSynced`, …) that `tests/common/mod.rs` also holds — `common` calls
itself a "generated copy" in its first line, but nothing generates it, so the
two were hand-synced and had drifted: this copy gained commonware-simplex
support (`build_cw`, the `app_cw` flag, `CwOut` → `CwIn` in-memory routing)
that `common` never received, and `common` had gained `InMemoryNetwork` and
`stamp_app_frame_output` that this copy lacked.

Merge the CW support into `tests/common/mod.rs` — now a true superset — and
have `e2e_consensus.rs` use it via `mod common`, deleting the local copy.
Each file under tests/ is its own crate, so this is the only way two test
binaries can share code; the rig cannot move to `quil_engine::test_support`
because it depends on dev-only features (`quil-execution/testing-stubs`,
`quil-hypergraph/test-utils`, parking_lot) that the lib build doesn't have.

e2e_consensus.rs: 3129 -> 1930 lines. No test bodies changed. Full suite
green: 2886 passed, 0 failed.

`materialize_harness.rs` keeps its own 5-line `materialize_synced` free
function rather than pulling the whole rig in for one helper.
